+++ src/test/java/kitchenpos/tablegroup/application/TableGroupServiceTest.java
@@ -13,9 +13,8 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderTable;
-import kitchenpos.tablegroup.application.TableGroupService;
 import kitchenpos.tablegroup.domain.TableGroup;
-import kitchenpos.order.dto.OrderTableInTableGroupDto;
+import kitchenpos.tablegroup.dto.OrderTableInTableGroupDto;
 import kitchenpos.tablegroup.dto.TableGroupCreateRequest;
 import kitchenpos.menu.repository.MenuGroupRepository;
 import kitchenpos.menu.repository.MenuRepository;
